robot/python/requirements.txt
ultralytics
opencv-python
pyyaml
pyserial
numpy
